Crate fc_storage

Structs

OverrideHandle

RuntimeApiStorageOverride

A wrapper type for the Runtime API. This type implements StorageOverride, so it can be used when calling the runtime API is desired but a dyn StorageOverride is required.

SchemaV1Override

An override for runtimes that use Schema V1

SchemaV2Override

An override for runtimes that use Schema V2

SchemaV3Override

An override for runtimes that use Schema V3

Traits

StorageOverride

Something that can fetch Ethereum-related data. This trait is quite similar to the runtime API, and indeed oe implementation of it uses the runtime API. Having this trait is useful because it allows optimized implementations that fetch data from a State Backend with some assumptions about pallet-ethereum’s storage schema. Using such an optimized implementation avoids spawning a runtime and the overhead associated with it.

Functions

onchain_storage_schema

overrides_handle