fc_storage

Module overrides

Structs

RuntimeApiStorageOverride

A storage override for runtimes that use runtime API.

SchemaV1StorageOverride

A storage override for runtimes that use schema v1.

SchemaV1StorageOverrideRef

A storage override reference for runtimes that use schema v1.

SchemaV2StorageOverride

A storage override for runtimes that use schema v2.

SchemaV2StorageOverrideRef

A storage override reference for runtimes that use schema v2.

SchemaV3StorageOverride

A storage override for runtimes that use schema v3.

SchemaV3StorageOverrideRef

A storage override for runtimes that use schema v3.

StorageQuerier

A useful utility for querying storage.

Traits

StorageOverride

This trait is used to obtain Ethereum-related data.